Enable transit->shamir seal migration in Enterprise
